Allow --out-dir customisation via rustc_args
When building a binary crate type, `cargo-lite build` puts the resulting artifact
in the crate root (`src/` in the common repository layout).

cargo-lite always included `--out-dir` when shelling out to `rustc`, so the user
could not override it with `rustc_args` in their `cargo-lite.conf` because the
compiler complained about setting `--out-dir` twice.

Now we only set `--out-dir` if it wasn't specified by the user.